add extra config files
chess-beast -- this is a homemade foot controller, it created from a the
circuit board of playstation 1 style pc gamepad, half a chessboard and
10 largish buttons. It is here more for archival purposes but if anybody
is interested I will do a tutorial on how to build one.

dancemat-overlap -- this is the dancemat with the buttons mapped
sequencially rather than by the button numbers - I am using this so I
can overlap loop ids between several devices.

nv61slmkii-1to8 -- this is a map for my novation controller keyboard. It
uses the 1 to 8 preset.